Generate adc connections

+++ rrrc/components/ADCDispatcher/ADCDispatcher.h
@@ -18,7 +18,7 @@ uint16_t ADCDispatcher_Read_ADC0_RawChannelData(uint32_t index);
 Voltage_t ADCDispatcher_Read_ADC0_ChannelVoltage(uint32_t index);
 uint16_t ADCDispatcher_Read_ADC1_RawChannelData(uint32_t index);
 Voltage_t ADCDispatcher_Read_ADC1_ChannelVoltage(uint32_t index);
-void ADCDispatcher_Write_Sensor_ADC(uint32_t index, uint16_t value);
+void ADCDispatcher_Write_Sensor_ADC(uint32_t index, uint8_t value);
 void ADCDispatcher_Write_MotorCurrent(uint32_t index, Current_t value);
 void ADCDispatcher_Write_MainBatteryVoltage(Voltage_t value);
 void ADCDispatcher_Write_MotorBatteryVoltage(Voltage_t value);
